Break profile editor/form into multiple part files and organize with nested directory
